% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{callS3methodParent}
\alias{callS3methodParent}
\title{Call the first method inhereted by the object}
\usage{
callS3methodParent(object, FUN, class = 1, ...)
}
\arguments{
\item{object}{an object}

\item{FUN}{the name of a generic method}

\item{class}{the level of class}

\item{...}{additional arguments to be passed to the method that will be called}
}
\description{
Call the first method inhereted by the object
}
\examples{
m <- lvm.reduced(Y~X)
callS3methodParent(m, "coef")
callS3methodParent(m, FUN = "coef", class = "lvm.reduced")

}

