% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_lp.R
\name{cancelLP}
\alias{cancelLP}
\alias{cancelLP.lvm.reduced}
\alias{cancelLP<-}
\alias{cancelLP<-.lvm.reduced}
\title{Remove variable from the linear predictor}
\usage{
cancelLP(x, ...)

cancelLP(x, ...) <- value

\method{cancelLP}{lvm.reduced}(x, ...) <- value

\method{cancelLP}{lvm.reduced}(x, link = NULL, lp = NULL, expar = TRUE,
  restaure = FALSE, simplify = TRUE, ...)
}
\arguments{
\item{x}{\code{lvm}-object}

\item{...}{argument passed to \code{cancelLP.lvm.reduced}}

\item{link, value}{the name of the links that should be removed}

\item{lp}{the name of the linear predictors that should be removed}

\item{expar}{should the external parameters be also removed}

\item{restaure}{should the link be kept while removing the linear predictor}

\item{simplify}{should the class \code{lvm.reduced} be removed is from the \code{lvm}-object if it contains no LP.}
}
\description{
Remove one or several variable from the linear predictor
}
\examples{
## regresssion
m <- lvm.reduced()
m <- regression(m, x=paste0("x",1:10),y="y", reduce = TRUE)
vars(m)

cancelLP(m)
cancelLP(m, restaure = TRUE)

lp.links <- lp(m, type = "link")
 newm <- cancelLP(m, lp = "LPy", link = lp.links[1:2])
lp(newm, type = "link")
}

