% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{convFormulaCharacter}
\alias{character2formula}
\alias{convFormulaCharacter}
\alias{formula2character}
\title{formula character conversion}
\usage{
formula2character(f, type = "formula")

character2formula(txt)
}
\arguments{
\item{f}{a formula.}

\item{type}{should the normal formula operator be used (\code{"formula"}) or the one of lava.option (\code{"symbols"} or \code{"symbol"}).}

\item{txt}{a character string.}
}
\description{
Conversion of formula into character string or vice versa
}
\examples{
formula2character(Y1~X1+X2)
formula2character(Y1~X1+X2, type = "symbols")

character2formula("Y1~X1+X2")

m <- lvm(Y~X)
character2formula(coef(m)[2])
}

