% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_extract.R, R/Reduce_object.R
\name{getReduce}
\alias{endogenous.lvm.reduced}
\alias{exogenous.lvm.reduced}
\alias{getReduce}
\alias{lp}
\alias{lp.lvm.reduced}
\alias{manifest.lvm.reduced}
\alias{reduce.lvm}
\alias{vars.lvm.reduced}
\title{Extract variables from a reduced latent variable model}
\usage{
lp(x, ...)

\method{lp}{lvm.reduced}(x, type = "name", lp = NULL, format = "vector",
  ...)

\method{vars}{lvm.reduced}(x, lp = TRUE, xlp = FALSE, ...)

\method{exogenous}{lvm.reduced}(x, lp = TRUE, xlp = FALSE, ...)

\method{endogenous}{lvm.reduced}(x, top = FALSE, latent = FALSE, ...)

\method{manifest}{lvm.reduced}(x, lp = TRUE, xlp = FALSE, ...)

\method{reduce}{lvm}(object, link = NULL, endo = NULL, rm.exo = TRUE)
}
\arguments{
\item{x}{\code{lvm}-object}

\item{type}{slot to be return. Can be \code{"link"}, \code{"x"}, \code{"con"}, \code{"name"},}

\item{lp}{should the name of the linear predictors be returned?}

\item{format}{should the results be kept as the list or returned as a single vector,}

\item{xlp}{should the name of the variables that the linear predictors aggregates be returned?}

\item{index}{which linear predictor to consider,}
}
\description{
Extract variables as in a standard lvm but including or not the variables that compose the linear predictor
}
\details{
lp returns all the linear predictors of the \code{lvm}-object. 
The other functions plays the same role as those defined in the lava package.
}
\examples{
 
## regression
m <- lvm.reduced()
m <- regression(m, x=paste0("x",1:10),y="y", reduce = TRUE)
vars(m)
vars(m, lp = TRUE)
vars(m, lp = FALSE, xlp = TRUE)

lp(m)

exogenous(m)
exogenous(m, lp = FALSE)
exogenous(m, xlp = TRUE)

endogenous(m)
endogenous(m, lp = FALSE) # should not change
endogenous(m, xlp = TRUE) # should not change

coef(m)

## lvm
m <- lvm.reduced()
m <- regression(m, x=paste0("x",1:10),y="y1", reduce = TRUE)
m <- regression(m, x=paste0("x",51:150),y="y2", reduce = TRUE)
covariance(m) <- y1~y2

vars(m)
vars(m, lp = FALSE)

lp(m)
lp(m, type = "x", format = "list")
lp(m, index = 1)

exogenous(m)
exogenous(m, lp = FALSE)

endogenous(m)
endogenous(m, lp = FALSE) # should not change
}

