% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{getS3methodParent}
\alias{getS3methodParent}
\title{Find methods belonging to other classes}
\usage{
getS3methodParent(object, FUN, class = 1, export = "all", value = FALSE)
}
\arguments{
\item{object}{an object}

\item{FUN}{the name of a generic method}

\item{class}{the level of class}

\item{export}{should all methods be output (\code{"all"}) or only the first one (\code{"first"})}

\item{value}{should the method be returned. Else its name will be returned.}
}
\description{
Find methods that apply to the object for each class of the object
}
\examples{
m <- lvm.reduced(Y~X)
getS3methodParent(m, "coef")
getS3methodParent(m, "coef", value = TRUE)

}

