% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_initCoef.R
\name{initCoef}
\alias{initCoef}
\title{Initialise \code{lvm.reduced}}
\usage{
initCoef(x, data, optim, ...)

initCoef(x, data, optim, ...)
}
\arguments{
\item{x}{\code{lvm.reduced}-object}

\item{data}{\code{data.frame}}

\item{optim}{control/optimization parameters}

\item{...}{additional arguments to be passed to the low level functions}
}
\description{
Estimate initial values for the parameters contained in a \code{lvm.reduced} object
}

