% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_initializer.R
\name{initializer}
\alias{initializer}
\alias{initializer.lava.reduce}
\title{Initialise \code{lvm.reduced}}
\usage{
initializer.lava.reduce(x, data, optim, ...)
}
\arguments{
\item{x}{\code{lvm.reduced}-object}

\item{data}{\code{data.frame}}

\item{optim}{control/optimization parameters}

\item{...}{additional arguments to be passed to the low level functions}
}
\description{
Estimate initial values for the parameters contained in a \code{lvm.reduced} object
}

