% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{initVar_link}
\alias{initVar_link}
\title{Normalize var1 and var2}
\usage{
initVar_link(var1, var2, repVar1 = FALSE, format = "list")
}
\arguments{
\item{var1}{a character indicating the endogeneous variable or a formula}

\item{var2}{an optional character indicating the exogeneous variable}

\item{repVar1}{should var1 be duplicated to match var2 length. Only active if format = "list".}

\item{format}{should the name of the variable be return (format = "list"), a vector of character formula ("txt.formula") or a list of formula ("formula")}
}
\description{
Convert var1 and var2 from formula or covariance to character
}
\details{
See test file test/testthat/test-Reduce.R for examples
}

