% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_kill.R
\name{kill.lp}
\alias{kill.lp}
\alias{lava.reduce.cancel.hook}
\alias{lava.reduce.remove.hook}
\title{Remove variables from linear predictors}
\usage{
lava.reduce.remove.hook(x, var, expar = TRUE, restaure = FALSE, ...)

lava.reduce.cancel.hook(x, value, expar = TRUE, restaure = FALSE, ...)
}
\arguments{
\item{x}{\code{lvm}-object}

\item{var}{the names of the variables that should be removed}

\item{expar}{should the external parameters be also removed}

\item{restaure}{should the link be kept while removing the linear predictor}

\item{...}{argument passed to \code{clean}}

\item{value}{the names of the variables that should be removed}
}
\description{
Remove one or several variable from the linear predictors
}
\examples{

m <- lvm()
m <- regression(m, x=paste0("x",1:10),y="y")
kill(m) <- "x7"

rm <- reduce(m, rm.exo = FALSE)

kill(rm) <- ~x6
kill(rm, value = "x1")
kill(rm, value = c("x2","x3"))
kill(rm, value = ~x4+x5)
kill(rm, value = ~LPy)


m <- lvm.reduced()
regression(m) <- y1 ~ x1+x2+x3
regression(m) <- y2 ~ z
covariance(m) <- y1 ~ y2
cancel(m) <- y1~x1+x2+x3
m
cancel(m) <- "y1 ~~ y2"
coef(m)

# see test/testthat/test-cancel.R
}

