% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_latent.R
\name{latent<-.lvm.reduced}
\alias{latent<-.lvm.reduced}
\title{Specify the latent variables}
\usage{
\method{latent}{lvm.reduced}(x, ...) <- value
}
\arguments{
\item{x}{\code{lvm}-object}

\item{...}{additional arguments to be passed to the low level functions}

\item{value}{the names of the latent variables}
}
\description{
Same as latent for latent variable model but check the consistency of the object
}

