% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_hook.R
\name{lava.reduced.estimate.hook}
\alias{lava.reduced.estimate.hook}
\title{Hook to estimate a reduce lvm model}
\usage{
lava.reduced.estimate.hook(x, data, weight, weight2, estimator, ...)
}
\description{
Add LP to dataset, update the estimator for handling LP, and find initialisation.
}
\examples{
#### regression ####

## simulation
m <- lvm()
m <- regression(m,y='y1',x='x'\%++\%1:2)
m <- regression(m,y='y1',x='z'\%++\%1:5)

set.seed(10)
d <- sim(m,150)
  
## reduced model 1
mR1 <- lvm.reduced()
mR1 <- regression(mR1,y='y1',x='x'\%++\%1:2)
mR1 <- regression(mR1,y='y1',x='z'\%++\%1:5, reduce = TRUE)

## reduced model 2
mR2 <- reduce(m)

## check estimation
em1 <- estimate(m, d, estimator = "gaussian1", control = list(trace = 2))
emR1 <- estimate(mR1, d, estimator = "gaussian1", control = list(trace = 2))
coef(em1) - coef(emR1)[names(coef(em1))]

#### penalization ####
mp <- m
penalize(mp) <- grep("x", coef(m),value = TRUE)
rmp <- reduce(mp)

epm <- estimate(mp,d,lambda1 = 1e5, control = list (trace = 2, iter.max = 2000))
epmR <- estimate(rmp,d,lambda1 = 1e5, control = list (trace = 2, iter.max = 2000))
coef(epm) - coef(epmR)[names(coef(epm))]


#### latent variable model ####
m <- lvm()
m <- regression(m,y=c('y1','y2','y3','y4'),x='eta')
m <- regression(m,y=c('y2','y3'),x='x'\%++\%1:5)
latent(m) <- ~eta
m <- regression(m,y=c('y1','y2'),x='z'\%++\%1:2)
covariance(m) <- y2~y1

# simul
set.seed(10)
d <- sim(m,100)

# reduced model
mR1 <- reduce(m, endo = c("y2"))
mR2 <- reduce(m)

## estimation
em <- estimate(m,d)
emR <- estimate(mR1, data = d, control = list(trace=2))

}

