% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lava.reduce-package.R
\docType{package}
\name{lavaReduce}
\alias{lavaReduce}
\alias{lavaReduce-package}
\title{lavaReduce}
\description{
Latent variable models with linear predictors. Should enable a faster optimisation when dealing with a large number of variables.
}
