% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_lp.R
\name{lp}
\alias{lp}
\alias{lp.lvm.reduced}
\title{Extract the linear predictors}
\usage{
lp(x, ...)

\method{lp}{lvm.reduced}(x, type = "name", lp = NULL, format = "vector",
  ...)
}
\arguments{
\item{x}{\code{lvm}-object}

\item{...}{additional arguments to be passed to the low level functions}

\item{type}{slot to be return. Can be \code{"link"}, \code{"x"}, \code{"con"}, \code{"name"}}

\item{lp}{which linear predictor to consider}

\item{format}{should the results be kept as the list or returned as a single vector}
}
\description{
Extract the linear predictors of a lvm object
}
\examples{
## regresssion
m <- lvm.reduced()
m <- regression(m, x=paste0("x",1:10),y="y", reduce = TRUE)
lp(m)
lp(m, type = c("x","link"), format = "list2")
lp(m, type = NULL)

## lvm
m <- lvm.reduced()
m <- regression(m, x=paste0("x",1:10),y="y1", reduce = TRUE)
m <- regression(m, x=paste0("x",51:150),y="y2", reduce = TRUE)
covariance(m) <- y1~y2

lp(m)
lp(m, type = "x", format = "list")
lp(m, lp = 1, type = "link")
}

