% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_object.R
\name{lvm.reduced}
\alias{lvm.reduced}
\title{Initialize new latent variable model with linear predictors}
\usage{
lvm.reduced(...)
}
\arguments{
\item{...}{arguments to be passed to the lvm function}
}
\description{
Function that constructs a new latent variable model object with linear predictors
}
\examples{
m <- lvm.reduced(Y ~ X1+X2)
class(m)
lp(m)

}

