% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_object.R
\name{lvm2reduce}
\alias{lvm2reduce}
\title{Conversion to a latent variable model with linear predictors}
\usage{
lvm2reduce(x)
}
\arguments{
\item{x}{\code{lvm}-object}
}
\description{
Add the possibility to use linear predictor in a latent variable model
}
