% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_moment.R
\docType{data}
\name{momentLVMr}
\alias{gaussian1LP_gradient.lvm}
\alias{gaussian1LP_hessian.lvm}
\alias{gaussian1LP_logLik.lvm}
\alias{gaussian1LP_objective.lvm}
\alias{gaussian1LP_score.lvm}
\alias{gaussian2LP_gradient.lvm}
\alias{gaussian2LP_hessian.lvm}
\alias{gaussian2LP_logLik.lvm}
\alias{gaussian2LP_objective.lvm}
\alias{gaussian2LP_score.lvm}
\alias{gaussianLP_gradient.lvm}
\alias{gaussianLP_hessian.lvm}
\alias{gaussianLP_logLik.lvm}
\alias{gaussianLP_method.lvm}
\alias{gaussianLP_objective.lvm}
\alias{gaussianLP_score.lvm}
\alias{momentLVMr}
\title{Moments for a reduced lvm model}
\format{An object of class \code{character} of length 1.}
\usage{
gaussianLP_method.lvm

gaussianLP_objective.lvm(x, p, data, ...)

gaussianLP_logLik.lvm(object, p, data, ...)

gaussianLP_gradient.lvm(x, p, data, ...)

gaussianLP_score.lvm(x, p, data, indiv = FALSE, ...)

gaussianLP_hessian.lvm(x, p, n, type, ...)

gaussian1LP_logLik.lvm(object, p, data, ...)

gaussian1LP_objective.lvm(x, p, data, ...)

gaussian1LP_score.lvm(x, p, data, indiv = FALSE, ...)

gaussian1LP_gradient.lvm(x, p, data, ...)

gaussian1LP_hessian.lvm(x, type, ...)

gaussian2LP_logLik.lvm(object, p, data, ...)

gaussian2LP_objective.lvm(x, p, data, ...)

gaussian2LP_score.lvm(x, p, data, indiv = FALSE, ...)

gaussian2LP_gradient.lvm(x, p, data, ...)

gaussian2LP_hessian.lvm(x, type, ...)
}
\arguments{
\item{x, object}{\code{lvm.reduced}-object}

\item{p}{current parameter estimate}

\item{data}{dataset}

\item{...}{additional arguments}

\item{indiv}{should the individual contribution be returned? Else average over the observations}

\item{n}{number of observations}

\item{type}{method used to compute the hessian}
}
\description{
Compute the likelihood, score and hessian of a reduced lvm model
}
\details{
this function assumes that the external parameters in p are at the end of the vector
}
\keyword{datasets}

