% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_moment.R
\name{moments.lvmfit.reduced}
\alias{moments.lvmfit.reduced}
\title{Extract moments from a reduced lvm}
\usage{
\method{moments}{lvmfit.reduced}(x, p = pars(x), ...)
}
\arguments{
\item{x}{latent variable model}

\item{p}{Parameter vector used to calculate statistics}

\item{...}{additional arguments to be passed to the low level functions}
}
\description{
Extract moments from a reduced lvm
}
