% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_object.R
\name{reduce}
\alias{reduce}
\alias{reduce.lvm}
\title{Reduce latent variable model}
\usage{
reduce(x, ...)

\method{reduce}{lvm}(x, link = NULL, endo = NULL, rm.exo = TRUE, ...)
}
\arguments{
\item{x}{\code{lvm}-object}

\item{...}{additional arguments to be passed to the low level functions}

\item{link}{the links that should be aggregated into linear predictors}

\item{endo}{the endogeneous variables for which the related exogeneous variables should be aggregated}

\item{rm.exo}{should the exogeneous variables be remove from the object}
}
\description{
Aggregate exogeneous variables into a linear predictor
}
\examples{
## regression
m <- lvm()
m <- regression(m, x=paste0("x",1:10),y="y")

m
reduce(m)
coef(reduce(m))
reduce(m, rm.exo = FALSE)

## lvm
m <- lvm()
m <- regression(m, x=paste0("x",1:10),y="y1")
m <- regression(m, x=paste0("x",11:30),y="y2")
covariance(m) <- y1~y2

m
reduce(m)
coef(reduce(m))
reduce(m, rm.exo = FALSE)

}

