% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_object.R
\name{reduce2lvm}
\alias{reduce2lvm}
\alias{reduce2lvm.lvm.reduced}
\alias{reduce2lvm.lvmfit.reduced}
\title{Remove all linear predictors from a lvm.reduce object}
\usage{
reduce2lvm(x, ...)

\method{reduce2lvm}{lvm.reduced}(x)

\method{reduce2lvm}{lvmfit.reduced}(x)
}
\arguments{
\item{x}{\code{lvm}-object}

\item{...}{not used.}
}
\description{
Remove all linear predictors from a lvm.reduce object. Restaure the links with the covariates.
}
\examples{
m <- lvm(Y~X1+X2+X3)
mR <- reduce(m)

reduce2lvm(mR)

m <- lvm(c(Y1~X1+X2+X3,Y2~Z1+X3))
mR <- reduce(m)
regression(mR) <- Y3~X1+X8
reduce2lvm(mR)
 
}
