% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_regression.R
\name{regression.lvm.reduced}
\alias{regression.lvm.reduced}
\title{Add regression association to a latent variable model with linear predictor}
\usage{
\method{regression}{lvm.reduced}(object = lvm(), to, from, y, x,
  reduce = FALSE, ...)
}
\arguments{
\item{to}{Character vector of outcome(s) or formula object.}

\item{from}{Character vector of predictor(s).}

\item{y}{Alias for 'to'}

\item{x}{\code{lvm}-object}

\item{reduce}{should the from variable be grouped into a linear predictor}

\item{x}{Alias for 'from'}
}
\description{
Same as regression for latent variable model with an argument reduce to introduce linear predictors
}
\examples{

m <- lvm.reduced()
m <- regression(m,y='y1',x='x'\%++\%1:2)
m <- regression(m,y='y1',x='z'\%++\%1:5, reduce = TRUE)

}

