% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_regression.R
\name{regression}
\alias{regression}
\alias{regression.lvm.reduced}
\alias{regression<-.lvm.reduced}
\title{Add regression association to a latent variable model with linear predictor}
\usage{
\method{regression}{lvm.reduced}(object = lvm.reduced(), ...) <- value

\method{regression}{lvm.reduced}(object = lvm.reduced(), to, from, y, x,
  reduce = FALSE, value, ...)
}
\arguments{
\item{object}{\code{lvm}-object}

\item{...}{additional arguments to be passed to \code{regression.lvm}}

\item{value}{A formula specifying the linear constraints or if \code{to=NULL} a \code{list} of parameter values.}

\item{to}{Character vector of outcome(s) or formula object.}

\item{from}{Character vector of predictor(s).}

\item{y}{Alias for 'to'}

\item{x}{Alias for 'from'}

\item{reduce}{should the from variable be grouped into a linear predictor}
}
\description{
Same as regression for latent variable model with an argument reduce to introduce linear predictors
}
\examples{

m <- lvm.reduced()
m <- regression(m,y='y1',x='x'\%++\%1:2)
m <- regression(m,y='y1',x='z'\%++\%1:5, reduce = TRUE)
m

m <- lvm.reduced()
regression(m) <- y1 ~ x1 + x2
regression(m, reduce = TRUE) <- y1 ~ x5 + x6 + x7
regression(m, reduce = TRUE) <- as.formula(paste0("y~",paste("x",1:5,collapse="+",sep="")))

m <- lvm.reduced()
regression(m, reduce = "LL") <- y1 ~ x5 + x6 + x7
m

}

