% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_lp.R
\name{updateLP}
\alias{lp<-}
\alias{lp<-.lvm.reduced}
\alias{updateLP}
\title{Update linear predictors}
\usage{
lp(x, ...) <- value

\method{lp}{lvm.reduced}(x, lp = NULL, ...) <- value
}
\arguments{
\item{x}{\code{lvm}-object}

\item{...}{additional arguments to be passed to the low level functions}

\item{value}{the value that will be allocated}

\item{lp}{the name of the linear predictors to be updated}
}
\description{
Update linear predictors of a lvm object
}
\examples{
## regresssion
m <- lvm.reduced()
m <- regression(m, x=paste0("x",1:10),y="y", reduce = TRUE)

newLP <- lp(m, type = NULL)[[1]]
newLP$link <- newLP$link[1:3]
newLP$con <- newLP$con[1:3]
newLP$x <- newLP$x[1:3]

lp(m, lp = 1) <- newLP
lp(m, type = NULL)


}

