% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_object.R
\name{cancelLP}
\alias{cancelLP}
\alias{cancelLP.lvm.reduced}
\alias{cancelLP<-}
\alias{cancelLP<-.lvm.reduced}
\title{Remove variable from the linear predictor}
\usage{
cancelLP(x, ...)

cancelLP(x, ...) <- value

\method{cancelLP}{lvm.reduced}(x, ...) <- value

\method{cancelLP}{lvm.reduced}(x, link = NULL, lp = NULL, expar = TRUE,
  simplify = TRUE)
}
\arguments{
\item{x}{\code{lvm}-object}

\item{lp}{should the name of the variables corresponding to the linear predictors be returned?}

\item{expar}{should the external parameters be also removed}
}
\description{
Remove one or several variable from the linear predictor
}
\examples{
## regresssion
m <- lvm.reduced()
m <- regression(m, x=paste0("x",1:10),y="y", reduce = TRUE)
vars(m)

cancelLP(m)
lp.links <- lp(m, type = "link")
 newm <- cancelLP(m, lp = "LPy", link = lp.links[1:2])
lp(newm, type = "link")

}

