% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_extract.R
\name{getReduce}
\alias{getReduce}
\alias{vars.lvm.reduced}
\alias{vars.lvmfit.reduced}
\alias{exogenous.lvm.reduced}
\alias{exogenous.lvmfit.reduced}
\alias{endogenous.lvm.reduced}
\alias{endogenous.lvmfit.reduced}
\alias{manifest.lvm.reduced}
\alias{manifest.lvmfit.reduced}
\title{Extract variables from a reduced latent variable model}
\usage{
\method{vars}{lvm.reduced}(x, lp = TRUE, xlp = FALSE, ...)

\method{vars}{lvmfit.reduced}(x, lp = FALSE, xlp = TRUE, ...)

\method{exogenous}{lvm.reduced}(x, lp = TRUE, xlp = FALSE, ...)

\method{exogenous}{lvmfit.reduced}(x, lp = FALSE, xlp = TRUE, ...)

\method{endogenous}{lvm.reduced}(x, top = FALSE, latent = FALSE, ...)

\method{endogenous}{lvmfit.reduced}(x, lp = FALSE, xlp = TRUE, ...)

\method{manifest}{lvm.reduced}(x, lp = TRUE, xlp = FALSE, ...)

\method{manifest}{lvmfit.reduced}(x, lp = FALSE, xlp = TRUE, ...)
}
\arguments{
\item{x}{\code{lvm}-object}

\item{lp}{should the name of the linear predictors be returned?}

\item{xlp}{should the name of the variables that the linear predictors aggregates be returned?}

\item{...}{additional arguments to be passed to the low level functions}

\item{top}{for compatibility with endogenous.lvm}

\item{latent}{logical defining whether latent variables without parents should be included in the result}
}
\description{
Extract variables as in a standard lvm but including or not the variables that compose the linear predictor
}
\details{
lp returns all the linear predictors of the \code{lvm}-object. 
The other functions plays the same role as those defined in the lava package.
}
\examples{
 
## regression
m <- lvm.reduced()
m <- regression(m, x=paste0("x",1:10),y="y", reduce = TRUE)
vars(m)
vars(m, lp = TRUE)
vars(m, lp = FALSE, xlp = TRUE)

exogenous(m)
exogenous(m, lp = FALSE)
exogenous(m, xlp = TRUE)

endogenous(m)
endogenous(m, lp = FALSE) # should not change
endogenous(m, xlp = TRUE) # should not change

coef(m)

## lvm
m <- lvm.reduced()
m <- regression(m, x=paste0("x",1:10),y="y1", reduce = TRUE)
m <- regression(m, x=paste0("x",51:150),y="y2", reduce = TRUE)
covariance(m) <- y1~y2

vars(m)
vars(m, lp = FALSE)

exogenous(m)
exogenous(m, lp = FALSE)

endogenous(m)
endogenous(m, lp = FALSE) # should not change
}
