% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_onLoad.R
\name{hook.reduce}
\alias{addhook_lava.reduce}
\alias{gethook_lava.reduce}
\alias{hook.reduce}
\title{Hooks for lava reduce}
\usage{
gethook_lava.reduce(hook, ...)

addhook_lava.reduce(x, hook, ...)
}
\arguments{
\item{hook}{the name of the hook}

\item{...}{for compatibility with lava}

\item{x}{the function to add to the hook}
}
\description{
Get and add hook for lava reduce
}

