% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_hook.R
\name{hookLVMreduced}
\alias{hookLVMreduced}
\alias{lava.reduce.estimate.hook}
\alias{lava.reduce.post.hook}
\title{Hook for \code{lvm.reduced} objects}
\usage{
lava.reduce.estimate.hook(x, data, estimator, ...)

lava.reduce.post.hook(x)
}
\arguments{
\item{x}{\code{lvm.reduced}-object}

\item{data}{\code{data.frame}}

\item{estimator}{the function used to compute the derivatives of the objective function}

\item{...}{additional arguments to be passed to the low level functions}
}
\description{
Pre and post processing methods
}

