% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_kill.R
\name{killLP}
\alias{cancel.lvm.reduced}
\alias{kill.lvm.reduced}
\alias{killLP}
\title{Remove variables from linear predictors}
\usage{
\method{kill}{lvm.reduced}(x, value = NULL, lp = TRUE, expar = TRUE,
  restaure = FALSE, simplify = TRUE, ...)

\method{cancel}{lvm.reduced}(x, value = NULL, expar = TRUE,
  restaure = FALSE, simplify = TRUE, ...)
}
\arguments{
\item{x}{\code{lvm}-object}

\item{value}{the name of the links that should be removed}

\item{lp}{should the variables be removed from the linear predictor}

\item{expar}{should the external parameters be also removed}

\item{restaure}{should the link be kept while removing the linear predictor}

\item{simplify}{should the class \code{lvm.reduced} be removed is from the \code{lvm}-object if it contains no LP.}

\item{...}{argument passed to \code{kill.lvm.reduced}}
}
\description{
Remove one or several variable from the linear predictors
}
\examples{
# see test/testthat/test-cancel.R
}

