% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_object.R
\name{reduce}
\alias{reduce}
\alias{reduce.lvm}
\title{Reduce latent variable model}
\usage{
reduce(x, ...)

\method{reduce}{lvm}(x, link = NULL, endo = NULL, clean = TRUE, ...)
}
\arguments{
\item{x}{\code{lvm}-object}

\item{...}{additional arguments to be passed to the low level functions}

\item{link}{the links that should be aggregated into linear predictors}

\item{endo}{the endogeneous variables for which the related exogeneous variables should be aggregated}

\item{clean}{should the lvm object be simplified using the \code{clean} function}
}
\description{
Aggregate exogeneous variables into a linear predictor
}
\examples{
## regression
m <- lvm()
m <- regression(m, x=paste0("x",1:10),y="y")
m

mtest <- m
cancel(mtest) <- y ~ x1 + x2 + x3 + x4 + x5
mtest
coef(mtest)
rm <- reduce(m)
rm
lp(rm, type = "link")
coef(rm)

coef(reduce(m))
reduce(m, rm.exo = TRUE)

reduce(m, link = paste0("y~x",1:5))

## lvm
m <- lvm()
m <- regression(m, x=paste0("x",1:10),y="y1")
m <- regression(m, x=paste0("x",11:30),y="y2")
covariance(m) <- y1~y2

system.time(
rm <- reduce(m)
)
coef(rm)
reduce(m, rm.exo = TRUE)

}
